Changed hero image to a background image
The original implementation of the image did not allow for the more
complex scaling behavior necessary to insure support for a wide range
of displays. As a background, it can now scale horizontally and crop
vertically in order to adjust its aspect ratio dynamically without
distortion. It also avoids moving the main subject of the painting out
of the corner, which is an important element of the joke.