Change alignment/scaling of hero background image
The background image in the hero scaled adequately within a small range
of aspect ratios, but left unsightly blank space on narrow displays.
The new alignment of horizontally centered and vertically placed at the
bottom, combined with scaling that displays the full height of the
image and crops horizontally ensures a wider range of supported aspect
ratios.