Edit writing
Fixed various spelling errors, sections that read awkwardly. Changed
name of fictional dystopian company from Torment Nexus to perdit.io
because Torment Nexus is not an original name. It was a name in a joke
which was one of the inspirations for the creative writing aspect of
this exercise. It should not have been included given the public
visibility of this repo.