Change scaling behavior of hero img
Finding a way to ensure the scaling of the image maintained all desired
properties turned out to be less helpful as an opportunity for
learning, than it was unhelpful as a distraction from focusing on the
fundamentals. The image now scales horizontally and crops the bottom as
the aspect ratio varies. The joke of the screaming figure being pushed
out of frame is still present, but less obvious and consistent.